third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/styles-3/styles-commit-editing-expected.txt

Tests that editing is canceled properly after incremental editing.


Running: selectInitialNode

Running: testFreeFlowEdit
[expanded] 
element.style { ()
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;


Running: testCommitEditing

Running: testNewPropertyEditorIsCreated

Running: testCycleThroughPropertyEditing

Running: testNodeStyles

Running: dumpStyles
After append:
[expanded] 
element.style { ()
    color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;
